Experiment with consolidating

+++ backend/src/services/posts.js
@@ -1,6 +1,8 @@
 import { Post } from '../db/models/post.js'
- export async function createPost({ title, author, contents, tags }) { 
-    const post = new Post({ title, author, contents, tags }) 
+import { User } from '../db/models/user.js'
+
+ export async function createPost( userID, {title, contents, tags}={}) { 
+    const post = new Post({ title, author: userID, contents, tags }) 
     return await post.save() 
 }
 
@@ -14,8 +16,10 @@ export async function listAllPosts(option){
     return await listPosts({}, option)
 }
 
-export async function listPostsByAuthor(author, option) {
-    return await listPosts({ author}, option) 
+export async function listPostsByAuthor(authorUsername, option) {
+    const user = await User.findOne({ username: authorUsername })
+  if (!user) return []
+  return await listPosts({ author: user._id }, option)
 }
 
 export async function listPostsByTag(tags, option) {
@@ -26,11 +30,11 @@ export async function getPostById(postId) {
     return await Post.findById(postId) 
 }
 
-export async function updatePost(postId, { title, author, contents, tags }) { 
-    return await Post.findOneAndUpdate( { _id: postId }, { $set: { title, author, contents, tags } },
+export async function updatePost(userID, postId, { title, contents, tags }={}) { 
+    return await Post.findOneAndUpdate( { _id: postId, author:userID }, { $set: { title, contents, tags } },
         { new: true }, ) 
     }
 
-export async function deletePost(postId) { 
-    return await Post.deleteOne({ _id: postId }) 
+export async function deletePost(userID, postId) { 
+    return await Post.deleteOne({ _id: postId, author: userID }) 
 }
